refactor: GoalCardコンポーネントをGoalsPageから分離
GoalCard（約150行）を独立コンポーネントに分離し、
CATEGORIES定数・ヘルパー関数をGoalCard側に内包。
GoalsPageは463行→269行に削減、props渡しも簡素化。
